{{cookiecutter.project_slug}}/backend/requirements.txt
astroid==2.0.4
beautifulsoup4==4.6.3
celery==4.2.1
Django==2.1.2
channels==2.1.3
factory_boy==2.11.1
django-cors-headers==2.4.0
django-db-geventpool==3.0.0
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-floppyforms==1.7.0
django-hijack==2.1.10
django-htmlmin==0.10.0
django-parler==1.9.2
django-polymorphic==2.0.3
django-storages==1.7.1
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.3
gevent==1.3.7
isort==4.3.4
Pillow==5.3.0
psycopg2-binary==2.7.5
pylint==2.1.1
pylint-django==2.0.2
pytest-cov==2.6.0
pytest-django==3.4.3
python-memcached==1.59
snapshottest==0.5.0
splinter[django]==0.9.0
whitenoise==4.1
